Moved invokeActionFunction() test to ToolManager and updated references.
